# fraud_Guard/dependencies/FRAUDGUARD_API_OPENAPI.yaml
openapi: 3.0.0
info:
  title: FraudGuard Academy API
  description: |
    RESTful API for the FraudGuard Learn-to-Earn fraud prevention platform.
    
    **Authentication:** All protected endpoints require JWT Bearer token in Authorization header.
    
    **Base URL:** `http://localhost:8080/api/v1` (local dev)
    
    **Production:** `https://api.fraudguard.com/api/v1`
  version: 1.0.0
  contact:
    name: FraudGuard Support
    email: support@fraudguard.com
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development
  - url: https://staging-api.fraudguard.com/api/v1
    description: Staging Environment
  - url: https://api.fraudguard.com/api/v1
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
    User:
      type: object
      description: User profile
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        username:
          type: string
          minLength: 3
          maxLength: 20
          example: "john_doe"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        region:
          type: string
          description: User's region for leaderboard filtering
          enum: ["Western Cape", "Eastern Cape", "Northern Cape", "Free State", "KwaZulu-Natal", "Gauteng", "Limpopo", "Mpumalanga", "North West"]
          example: "Western Cape"
        phoneNumber:
          type: string
          pattern: ^\+?[0-9]{10,15}$
          example: "+27821234567"
        totalScore:
          type: integer
          description: Total cumulative score across all scenarios
          example: 4250
        currentStreak:
          type: integer
          description: Number of consecutive correct answers
          example: 7
        maxStreak:
          type: integer
          description: Highest streak achieved
          example: 23
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-20T14:45:00Z"
      required:
        - id
        - email
        - username
        - region
        - totalScore
    AuthRequest:
      type: object
      description: Login credentials
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          format: password
          example: "SecurePass123!"
      required:
        - email
        - password
    RegisterRequest:
      type: object
      description: User registration data
      properties:
        email:
          type: string
          format: email
          example: "newuser@example.com"
        username:
          type: string
          minLength: 3
          maxLength: 20
          example: "jane_smith"
        password:
          type: string
          minLength: 8
          format: password
          example: "SecurePass123!"
        firstName:
          type: string
          example: "Jane"
        lastName:
          type: string
          example: "Smith"
        region:
          type: string
          enum: ["Western Cape", "Eastern Cape", "Northern Cape", "Free State", "KwaZulu-Natal", "Gauteng", "Limpopo", "Mpumalanga", "North West"]
          example: "Western Cape"
        phoneNumber:
          type: string
          pattern: ^\+?[0-9]{10,15}$
          example: "+27821234567"
      required:
        - email
        - username
        - password
        - firstName
        - lastName
        - region
    AuthResponse:
      type: object
      description: JWT tokens returned after successful authentication
      properties:
        accessToken:
          type: string
          description: Short-lived JWT (valid 15 minutes)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          description: Long-lived refresh token (valid 7 days)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'
      required:
        - accessToken
        - refreshToken
        - user
    Scenario:
      type: object
      description: Fraud scenario learning content
      properties:
        id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          example: "WhatsApp Business Scam"
        description:
          type: string
          example: "A message appears in your WhatsApp from someone claiming to be your bank..."
        type:
          type: string
          enum: ["PHISHING_EMAIL", "FAKE_BANKING_APP", "WHATSAPP_SCAM", "SMS_PHISHING", "VISHING_CALL"]
          description: Category of fraud
          example: "WHATSAPP_SCAM"
        channel:
          type: string
          enum: ["MOBILE_APP", "USSD", "SMS"]
          description: Where scenario is presented
          example: "MOBILE_APP"
        difficulty:
          type: string
          enum: ["BEGINNER", "INTERMEDIATE", "ADVANCED"]
          example: "BEGINNER"
        imageUrl:
          type: string
          format: uri
          description: Image showing the fraudulent scenario (for mobile app)
          example: "https://cdn.fraudguard.com/scenarios/whatsapp-scam.jpg"
        hotspots:
          type: array
          description: Clickable UI areas (for interactive hotspot learning)
          items:
            type: object
            properties:
              id:
                type: integer
              label:
                type: string
              isCorrect:
                type: boolean
              explanation:
                type: string
          example:
            - id: 1
              label: "Sender's verified checkmark"
              isCorrect: false
              explanation: "Scammers can fake verified badges. Look for official bank contact instead."
            - id: 2
              label: "Urgent tone & security threat"
              isCorrect: true
              explanation: "Correct! Banks create urgency to panic you into action."
        correctAnswerId:
          type: integer
          description: ID of correct hotspot (for validation)
          example: 2
        explanation:
          type: string
          description: Educational explanation of why the correct answer is right
          example: "Banks never ask for OTPs via WhatsApp. Always call official support number."
        rewardAmount:
          type: number
          format: float
          description: Airtime reward for correct answer (in USD)
          example: 0.10
        createdAt:
          type: string
          format: date-time
          example: "2024-01-10T08:00:00Z"
      required:
        - id
        - title
        - description
        - type
        - channel
        - difficulty
        - correctAnswerId
        - explanation
        - rewardAmount
    ScenarioAnswer:
      type: object
      description: User's answer to a scenario
      properties:
        scenarioId:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440000"
        selectedHotspotId:
          type: integer
          description: ID of hotspot user clicked
          example: 2
      required:
        - scenarioId
        - selectedHotspotId
    ScenarioResult:
      type: object
      description: Result of scenario answer validation
      properties:
        isCorrect:
          type: boolean
          example: true
        explanation:
          type: string
          example: "Correct! Banks never ask for OTPs via WhatsApp."
        rewardEarned:
          type: number
          format: float
          description: Amount earned (0 if incorrect)
          example: 0.10
        streakUpdated:
          type: integer
          description: Current streak after this answer
          example: 8
        totalScore:
          type: integer
          description: Updated total score
          example: 4350
      required:
        - isCorrect
        - explanation
        - rewardEarned
    Wallet:
      type: object
      description: User's reward wallet
      properties:
        userId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        balanceUSD:
          type: number
          format: float
          description: Current balance in USD
          example: 2.50
        totalEarned:
          type: number
          format: float
          description: Total airtime earned (lifetime)
          example: 15.75
        totalRedeemed:
          type: number
          format: float
          description: Total airtime redeemed/paid out
          example: 13.25
        lastUpdated:
          type: string
          format: date-time
          example: "2024-01-20T14:45:00Z"
      required:
        - userId
        - balanceUSD
        - totalEarned
        - totalRedeemed
    PayoutRequest:
      type: object
      description: Request to redeem wallet balance as airtime
      properties:
        amount:
          type: number
          format: float
          minimum: 0.10
          maximum: 50.00
          description: Amount in USD to redeem (minimum $0.10)
          example: 2.50
        phoneNumber:
          type: string
          pattern: ^\+?[0-9]{10,15}$
          description: Phone number to send airtime to
          example: "+27821234567"
        provider:
          type: string
          enum: ["MTN", "VODACOM", "CELL_C", "TELKOM"]
          description: Mobile network provider
          example: "VODACOM"
      required:
        - amount
        - phoneNumber
        - provider
    PayoutResponse:
      type: object
      description: Payout request confirmation
      properties:
        payoutId:
          type: string
          format: uuid
          example: "770e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: ["PENDING", "PROCESSING", "COMPLETED", "FAILED"]
          example: "PROCESSING"
        amount:
          type: number
          format: float
          example: 2.50
        phoneNumber:
          type: string
          example: "+27821234567"
        provider:
          type: string
          example: "VODACOM"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-20T14:50:00Z"
        estimatedCompletion:
          type: string
          format: date-time
          description: Estimated time airtime will arrive
          example: "2024-01-20T15:00:00Z"
      required:
        - payoutId
        - status
        - amount
        - phoneNumber
    Transaction:
      type: object
      description: Wallet transaction (earning or redemption)
      properties:
        id:
          type: string
          format: uuid
          example: "880e8400-e29b-41d4-a716-446655440000"
        userId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        type:
          type: string
          enum: ["EARNED", "REDEEMED", "REVERSED"]
          example: "EARNED"
        amount:
          type: number
          format: float
          example: 0.10
        relatedScenarioId:
          type: string
          format: uuid
          description: Scenario ID if transaction is from earning
          example: "660e8400-e29b-41d4-a716-446655440000"
        relatedPayoutId:
          type: string
          format: uuid
          description: Payout ID if transaction is redemption
          example: "770e8400-e29b-41d4-a716-446655440000"
        description:
          type: string
          example: "Earned from WhatsApp Business Scam scenario"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-20T14:45:00Z"
      required:
        - id
        - userId
        - type
        - amount
        - description
        - createdAt
...
    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        error:
          type: string
          example: "INVALID_CREDENTIALS"
        message:
          type: string
          example: "Email or password is incorrect"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-20T14:50:00Z"
        path:
          type: string
          example: "/api/v1/auth/login"
      required:
        - error
        - message
security:
  - BearerAuth: []
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation failed (email already exists, weak password, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login with email and password
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token using refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              required:
                - refreshToken
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/profile:
    get:
      tags:
        - User
      summary: Get current user profile
      operationId: getUserProfile
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - User
      summary: Update user profile
      operationId: updateUserProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                region:
                  type: string
                  enum: ["Western Cape", "Eastern Cape", "Northern Cape", "Free State", "KwaZulu-Natal", "Gauteng", "Limpopo", "Mpumalanga", "North West"]
                phoneNumber:
                  type: string
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /scenarios:
    get:
      tags:
        - Scenarios
      summary: Get list of available scenarios (paginated)
      operationId: getScenarios
      parameters:
        - name: page
          in: query
          description: Page number (0-indexed)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Number of scenarios per page
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
        - name: difficulty
          in: query
          description: Filter by difficulty level
          schema:
            type: string
            enum: ["BEGINNER", "INTERMEDIATE", "ADVANCED"]
        - name: type
          in: query
          description: Filter by scenario type
          schema:
            type: string
            enum: ["PHISHING_EMAIL", "FAKE_BANKING_APP", "WHATSAPP_SCAM", "SMS_PHISHING", "VISHING_CALL"]
      responses:
        '200':
          description: Scenarios retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Scenario'
                  totalElements:
                    type: integer
                    example: 25
                  totalPages:
                    type: integer
                    example: 3
                  currentPage:
                    type: integer
                    example: 0
  /scenarios/{scenarioId}:
    get:
      tags:
        - Scenarios
      summary: Get single scenario details
      operationId: getScenario
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Scenario details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '404':
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /scenarios/{scenarioId}/answer:
    post:
      tags:
        - Scenarios
      summary: Submit answer to a scenario
      operationId: submitScenarioAnswer
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioAnswer'
      responses:
        '200':
          description: Answer validated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioResult'
        '400':
          description: Invalid answer format or scenario already answered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wallet:
    get:
      tags:
        - Wallet
      summary: Get user's wallet balance and transaction history
      operationId: getWallet
      parameters:
        - name: limit
          in: query
          description: Number of recent transactions to return
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Wallet retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  wallet:
                    $ref: '#/components/schemas/Wallet'
                  recentTransactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
  /wallet/payout:
    post:
      tags:
        - Wallet
      summary: Request airtime payout
      operationId: requestPayout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayoutRequest'
      responses:
        '202':
          description: Payout request accepted and queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutResponse'
        '400':
          description: Invalid request (insufficient balance, invalid amount, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many payout requests (rate limited)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...